drivers: wifi: uwp: Rename all c and h files
Drop prefix wifi_ of all c and h files.
Drop one parameter 'ctx_id' of hw_test.
Rename ssid_length.
Rename psk_length.
