+++ drivers/wifi/uwp/CMakeLists.txt
@@ -2,10 +2,10 @@ if(CONFIG_WIFI_UWP)
   zephyr_include_directories(./)
 
   zephyr_sources(
-	wifi_cmdevt.c
-	wifi_txrx.c
-	wifi_ipc.c
-	wifi_main.c
-	wifi_mem.c
+	cmdevt.c
+	txrx.c
+	ipc.c
+	main.c
+	mem.c
     )
 endif()
